+++ packages/datadog_flutter_plugin/lib/src/helpers.dart
@@ -142,3 +142,9 @@ InvalidAttributeInfo? findInvalidAttribute(Map<String, Object?> attributes,
     [String parentPropertyName = '']) {
   return _checkInvalidValue(attributes, parentPropertyName);
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` to support older versions of Flutter.
+T? ambiguate<T>(T? value) => value;